refactor(ui): Extract MetricsTableHeader
