+++ packages/ui/src/components/metrics-table-header/index.ts
@@ -1,0 +1,1 @@
+export * from './metrics-table-header';
