+++ packages/ui/src/components/metrics-table-header/metrics-table-header.i18n.ts
@@ -1,0 +1,6 @@
+export const BASELINE_TITLE = 'Baseline';
+export const CURRENT_TITLE = 'Current';
+
+export const SORT_VALUE = 'value';
+export const SORT_CHANGE = 'change';
+export const SORT_PERCENTUAL_CHANGE = 'percentual change';
