+++ packages/ui/src/components/metrics-table-header/metrics-table-header.module.css
@@ -1,0 +1,44 @@
+/* columns */
+.root .metricName {
+  width: 55%;
+}
+
+.root .job {
+  text-align: right;
+}
+
+.root .value {
+  text-align: right;
+  vertical-align: baseline;
+  width: 15%;
+}
+
+.root .delta {
+  text-align: right;
+  font-size: var(--size-xsmall);
+  vertical-align: baseline;
+  width: 10%;
+  color: pink;
+}
+
+.root .deltaPercentage {
+  width: 5%;
+}
+
+.jobName {
+  white-space: nowrap;
+}
+
+/* columns overrides for showHeaderSum */
+.showHeaderSum .headerRow .job {
+  border-bottom: 0;
+  padding-bottom: 0;
+}
+
+.showHeaderSum .headerRow .sum {
+  padding-top: var(--space-xxxsmall);
+  font-size: var(--size-small);
+  vertical-align: middle;
+  text-transform: none;
+}
+
