+++ packages/ui/src/components/metrics-table/metrics-table.i18n.ts
@@ -1,10 +1,3 @@
-export const BASELINE_TITLE = 'Baseline';
-export const CURRENT_TITLE = 'Current';
-
-export const SORT_VALUE = 'value';
-export const SORT_CHANGE = 'change';
-export const SORT_PERCENTUAL_CHANGE = 'percentual change';
-
 export const EMPTY_MESSAGE = 'No items found.';
 export const SHOW_ALL = 'Show all';
 export const SHOW_LESS = 'Show less';
